+++ generators/server/templates/src/boot.js
@@ -2,5 +2,6 @@ import "./boot/environments/development.js";
 import "./boot/environments/test.js";
 
 <% if(options["db-client"] === "objection") { %>
-import "./boot/model.js";
+import connection from "./boot/model.js";
+export { connection }
 <% } %>