+++ generators/server/templates/src/models/index.js
@@ -1,0 +1,3 @@
+// include all of your models here using CommonJS requires
+
+module.exports = {};